+++ modules/edhoc/src/cose.c
@@ -32,7 +32,7 @@ enum err cose_enc_structure_encode(const uint8_t *context, uint32_t context_len,
 	enc_structure._enc_structure_external_aad.value = external_aad;
 	enc_structure._enc_structure_external_aad.len = external_aad_len;
 
-	uint32_t payload_len_out;
+	size_t payload_len_out;
 	TRY_EXPECT(cbor_encode_enc_structure(out, *out_len, &enc_structure,
 					     &payload_len_out),
 		   true);
@@ -59,7 +59,7 @@ enum err cose_sig_structure_encode(const uint8_t *context, uint32_t context_len,
 	sig_structure._sig_structure_payload.value = payload;
 	sig_structure._sig_structure_payload.len = payload_len;
 
-	uint32_t payload_len_out;
+	size_t payload_len_out;
 	TRY_EXPECT(cbor_encode_sig_structure(out, *out_len, &sig_structure,
 					     &payload_len_out),
 		   true);
